+++ src/routes/index.ts
@@ -1,0 +1,2 @@
+export { router as aliceRouter } from './alice-api';
+export { router as userRouter } from './user-api';
