+++ src/server.ts
@@ -8,6 +8,5 @@ mqtt.initMQTT();
 /* Activate the sever */
 
 app.listen(PORT, () => {
-  // tslint:disable-next-line:no-console
   console.log(`🌲 [server]: Pino is running at http://localhost:${PORT}`);
 });
